requirements.txt
hyperliquid-python-sdk>=0.10.0
python-dotenv>=1.0.0
keyring>=24.0.0
streamlit>=1.40.0
plotly>=5.24.0
pandas>=2.2.0
